requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.19
botocore==1.24.19
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
install==1.3.5
jmespath==0.10.0
Pillow==9.0.1
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
